fix: hit-test passthrough for AbsoluteFill, intrinsic measurement
Root cause:
1. PaintOps were incorrectly blocking hit-testing, causing transparent
   overlays to consume input intended for background interactables.
2. Text nodes lacked intrinsic measurement during lowering, causing
   flexible containers to collapse to 0x0 before the layout pass.

Fixes:
- Updated hit_test.rs to only consider Semantics, Scroll, and Embed ops
  as valid hit targets.
- Integrated TextMeasurer into Env and LoweringContext.
- build_layout_tree now performs intrinsic text measurement to set
  correct base constraints for flexible nodes.

Verified:
1. editor_qa screenshots show correctly laid out editor with content.
2. Interactive elements under overlays are now clickable.

+++ crates/core/fission-core/src/ui/widgets/lazy_column.rs
@@ -116,8 +116,8 @@ impl Lower for LazyColumn {
         let item_h = self.item_height.max(1.0);
         let total_count = self.children.len();
         
-        let start_index = (scroll_offset / item_h).floor() as usize;
-        let visible_count = (viewport_height / item_h).ceil() as usize + 1; // +1 buffer
+        let start_index = (scroll_offset / item_h as f32).floor() as usize;
+        let visible_count = (viewport_height as f32 / item_h as f32).ceil() as usize + 1; // +1 buffer
         let end_index = (start_index + visible_count).min(total_count);
         let start_index = start_index.min(total_count); // Clamp
